feat: remove test-matching, OverviewMapBanner, OverviewKpiTiles, riders/maintenance tabs
- git rm 12 dead files: app/test-matching/*, app/monitoring/*, app/overview/page.tsx,
  app/api/test-matching/export/[type]/route.ts,
  components/test-matching/{Matching,Rider,Vehicle}Section.tsx,
  components/overview/OverviewKpiTiles.tsx, components/overview/OverviewMapBanner.tsx
- app/page.tsx: drop MaintenancePanel/RidersPanel/CreateRiderDialog/CreateMaintenanceItemDialog
  imports; change InsuranceOption to type-only import; remove OverviewKpiTiles/OverviewMapBanner
  imports and JSX; trim TabKey to "vehicles"|"stations", TABS to 차량+충전소 (BSS→충전소);
  remove riders/maintenance activeContent branches; remove KPI computation vars
  (totalRiders, ignitionOnCount, insuredVehicleCount, subscriptionRiderCount,
  rentalRiderCount, imeiMinusOneSet, summary); remove formatCount/numberFormatter;
  trim tab action row to vehicles+stations dialogs only
- app/actions.ts: add setVehicleIgnitionBlockFromDashboardAction (migrated from
  deleted app/monitoring/actions.ts) for BikeDetailPanel ignition toggle
- components/dashboard/BikeDetailPanel.tsx: update import to use new dashboard action
