+++ .superpowers/brainstorm/291-1781141253/state/events
@@ -1,0 +1,1 @@
+{"type":"click","text":"A\n        \n          지도 클릭으로 핀\n          다이얼로그에 미니맵 — 클릭한 위치에 핀 꽂음. 주소는 텍스트로 직접 입력","choice":"map-click","id":null,"timestamp":1781142083496}
