+++ .superpowers/brainstorm/291-1781141253/state/server-info
@@ -1,0 +1,1 @@
+{"type":"server-started","port":63518,"host":"127.0.0.1","url_host":"localhost","url":"http://localhost:63518","screen_dir":"C:\\Users\\user\\repositories\\clever\\thundercrew-domain\\.superpowers\\brainstorm\\291-1781141253\\content","state_dir":"C:\\Users\\user\\repositories\\clever\\thundercrew-domain\\.superpowers\\brainstorm\\291-1781141253\\state"}
