+++ development/front-admin-web/.superpowers/brainstorm/1664-1779903457/content/dialog-layout.html
@@ -1,0 +1,54 @@
+<h2>VehicleDetailDialog — 다음 고객 섹션 위치</h2>
+<p class="subtitle">CLEANING 차량에만 표시되는 섹션입니다</p>
+
+<div class="split">
+  <div class="mockup">
+    <div class="mockup-header">A — 배송 섹션 위 (상단)</div>
+    <div class="mockup-body" style="font-size:12px;line-height:1.8;">
+      <div style="padding:6px 10px;background:#1e293b;border-bottom:1px solid #334155;font-weight:600;color:#94a3b8;font-size:10px;letter-spacing:.05em;">기본 정보</div>
+      <div style="padding:6px 10px;border-bottom:1px solid #1e293b;color:#e2e8f0;">차량번호 · 서비스유형 · 운영상태…</div>
+
+      <div style="padding:6px 10px;background:#0f4c2a;border-bottom:1px solid #166534;font-weight:600;color:#4ade80;font-size:10px;letter-spacing:.05em;">🧹 다음 고객 <span style="font-size:9px;color:#86efac;">(CLEANING 전용)</span></div>
+      <div style="padding:8px 10px;border-bottom:1px solid #1e293b;">
+        <div style="color:#94a3b8;font-size:10px;">고객 이름</div>
+        <div style="background:#1e293b;border-radius:4px;padding:4px 8px;margin-top:2px;color:#e2e8f0;">홍길동</div>
+        <div style="color:#94a3b8;font-size:10px;margin-top:6px;">전화번호</div>
+        <div style="background:#1e293b;border-radius:4px;padding:4px 8px;margin-top:2px;color:#e2e8f0;">010-1234-5678</div>
+        <div style="color:#94a3b8;font-size:10px;margin-top:6px;">주소</div>
+        <div style="background:#1e293b;border-radius:4px;padding:4px 8px;margin-top:2px;color:#e2e8f0;">서울 강남구 역삼동 123</div>
+        <button style="margin-top:8px;padding:3px 10px;background:#2563eb;color:#fff;border:none;border-radius:4px;font-size:11px;cursor:pointer;">저장</button>
+      </div>
+
+      <div style="padding:6px 10px;background:#1e293b;border-bottom:1px solid #334155;font-weight:600;color:#94a3b8;font-size:10px;letter-spacing:.05em;">배송 상태</div>
+      <div style="padding:6px 10px;border-bottom:1px solid #1e293b;color:#e2e8f0;">이동 중 · 진행률 42%…</div>
+
+      <div style="padding:6px 10px;background:#1e293b;border-bottom:1px solid #334155;font-weight:600;color:#94a3b8;font-size:10px;">텔레메트리 · 보험 · 정비…</div>
+    </div>
+  </div>
+
+  <div class="mockup">
+    <div class="mockup-header">B — 배송 섹션 안에 통합 (하단)</div>
+    <div class="mockup-body" style="font-size:12px;line-height:1.8;">
+      <div style="padding:6px 10px;background:#1e293b;border-bottom:1px solid #334155;font-weight:600;color:#94a3b8;font-size:10px;letter-spacing:.05em;">기본 정보</div>
+      <div style="padding:6px 10px;border-bottom:1px solid #1e293b;color:#e2e8f0;">차량번호 · 서비스유형 · 운영상태…</div>
+
+      <div style="padding:6px 10px;background:#1e293b;border-bottom:1px solid #334155;font-weight:600;color:#94a3b8;font-size:10px;">배송 상태</div>
+      <div style="padding:6px 10px;border-bottom:1px solid #1e293b;color:#e2e8f0;">이동 중 · 진행률 42%…</div>
+
+      <div style="padding:6px 10px;background:#0f4c2a;border-bottom:1px solid #166534;font-weight:600;color:#4ade80;font-size:10px;letter-spacing:.05em;">🧹 다음 고객 <span style="font-size:9px;color:#86efac;">(CLEANING 전용)</span></div>
+      <div style="padding:8px 10px;border-bottom:1px solid #1e293b;">
+        <div style="color:#94a3b8;font-size:10px;">고객 이름</div>
+        <div style="background:#1e293b;border-radius:4px;padding:4px 8px;margin-top:2px;color:#e2e8f0;">홍길동</div>
+        <div style="color:#94a3b8;font-size:10px;margin-top:6px;">전화번호</div>
+        <div style="background:#1e293b;border-radius:4px;padding:4px 8px;margin-top:2px;color:#e2e8f0;">010-1234-5678</div>
+        <div style="color:#94a3b8;font-size:10px;margin-top:6px;">주소</div>
+        <div style="background:#1e293b;border-radius:4px;padding:4px 8px;margin-top:2px;color:#e2e8f0;">서울 강남구 역삼동 123</div>
+        <button style="margin-top:8px;padding:3px 10px;background:#2563eb;color:#fff;border:none;border-radius:4px;font-size:11px;cursor:pointer;">저장</button>
+      </div>
+
+      <div style="padding:6px 10px;background:#1e293b;border-bottom:1px solid #334155;font-weight:600;color:#94a3b8;font-size:10px;">텔레메트리 · 보험 · 정비…</div>
+    </div>
+  </div>
+</div>
+
+<p class="subtitle" style="margin-top:16px;">A: 기본 정보 바로 아래, 눈에 잘 띔 &nbsp;|&nbsp; B: 배송 상태 아래, 맥락상 자연스러움</p>
