+++ development/front-admin-web/.superpowers/brainstorm/1664-1779903457/content/notification-design.html
@@ -1,0 +1,53 @@
+<h2>벨 알림 드롭다운 — 다음 고객 표시 방식</h2>
+<p class="subtitle">클리닝 차량 시동 ON 시 나타나는 알림 항목</p>
+
+<div class="split">
+  <div class="mockup">
+    <div class="mockup-header">A — 한 줄 요약</div>
+    <div class="mockup-body" style="font-family:sans-serif;">
+      <div style="padding:10px 14px;border-bottom:1px solid #1e293b;display:flex;justify-content:space-between;align-items:center;gap:8px;">
+        <span style="font-size:12px;color:#e2e8f0;font-weight:500;">📞 서울12가3456 → 홍길동 010-1234-5678</span>
+        <span style="font-size:11px;color:#64748b;white-space:nowrap;">방금</span>
+      </div>
+      <div style="padding:10px 14px;border-bottom:1px solid #1e293b;display:flex;justify-content:space-between;align-items:center;gap:8px;">
+        <span style="font-size:12px;color:#e2e8f0;font-weight:500;">📞 서울34나5678 → 김철수 010-9876-5432</span>
+        <span style="font-size:11px;color:#64748b;white-space:nowrap;">3분 전</span>
+      </div>
+    </div>
+  </div>
+
+  <div class="mockup">
+    <div class="mockup-header">B — 2줄 카드 (추천)</div>
+    <div class="mockup-body" style="font-family:sans-serif;">
+      <div style="padding:10px 14px;border-bottom:1px solid #1e293b;">
+        <div style="display:flex;justify-content:space-between;align-items:flex-start;">
+          <span style="font-size:12px;color:#4ade80;font-weight:600;">🧹 서울12가3456 출발</span>
+          <span style="font-size:11px;color:#64748b;white-space:nowrap;">방금</span>
+        </div>
+        <div style="margin-top:3px;font-size:11px;color:#94a3b8;">
+          📞 홍길동 &nbsp;·&nbsp; <span style="color:#60a5fa;">010-1234-5678</span>
+        </div>
+      </div>
+      <div style="padding:10px 14px;border-bottom:1px solid #1e293b;">
+        <div style="display:flex;justify-content:space-between;align-items:flex-start;">
+          <span style="font-size:12px;color:#4ade80;font-weight:600;">🧹 서울34나5678 출발</span>
+          <span style="font-size:11px;color:#64748b;white-space:nowrap;">3분 전</span>
+        </div>
+        <div style="margin-top:3px;font-size:11px;color:#94a3b8;">
+          📞 김철수 &nbsp;·&nbsp; <span style="color:#60a5fa;">010-9876-5432</span>
+        </div>
+      </div>
+    </div>
+  </div>
+</div>
+
+<div style="margin-top:20px;" class="mockup">
+  <div class="mockup-header">C — 고객 정보 없는 경우 (다음 고객 미설정)</div>
+  <div class="mockup-body" style="font-family:sans-serif;">
+    <div style="padding:10px 14px;display:flex;justify-content:space-between;align-items:center;gap:8px;">
+      <span style="font-size:12px;color:#e2e8f0;font-weight:500;">🔑 서울56다7890 이동 시작</span>
+      <span style="font-size:11px;color:#64748b;white-space:nowrap;">방금</span>
+    </div>
+    <p style="padding:0 14px 10px;font-size:11px;color:#64748b;margin:0;">※ 다음 고객 미설정 시 기존 알림 유지</p>
+  </div>
+</div>
