+++ development/front-admin-web/.superpowers/brainstorm/1664-1779903457/state/server-stopped
@@ -1,0 +1,1 @@
+{"reason":"idle timeout","timestamp":1779905918567}
