+++ development/front-admin-web/.superpowers/brainstorm/1755-1779775818/content/notification-style.html
@@ -1,0 +1,95 @@
+<h2>콜 알림 스타일</h2>
+<p class="subtitle">일정 등록 후 "콜 발송" 시 화면에 어떻게 나타나면 좋을까요?</p>
+
+<div class="cards">
+
+  <div class="card" data-choice="a" onclick="toggleSelect(this)">
+    <div class="card-image">
+      <div style="position:relative;font-family:sans-serif;font-size:12px;background:#f4f6f5;border-radius:8px;padding:12px;min-height:120px;border:1px solid #ddd;">
+        <div style="font-size:10px;color:#6b7280;margin-bottom:6px;">관리자 콘솔 화면...</div>
+        <!-- Toast -->
+        <div style="position:absolute;top:10px;right:10px;background:#111;color:#fff;border-radius:8px;padding:10px 14px;font-size:11px;min-width:200px;box-shadow:0 4px 12px rgba(0,0,0,0.3);">
+          <div style="display:flex;align-items:center;gap:8px;margin-bottom:4px;">
+            <span style="font-size:14px;">🔔</span>
+            <span style="font-weight:700;">콜 발송됨</span>
+          </div>
+          <div style="color:#a3a3a3;font-size:10px;">서울12가3456 → 강남구 역삼동</div>
+          <div style="color:#a3a3a3;font-size:10px;">05/27 10:00</div>
+          <div style="margin-top:6px;background:#3b82f6;border-radius:4px;padding:2px 0;text-align:center;font-size:10px;cursor:pointer;">확인</div>
+        </div>
+        <div style="font-size:10px;color:#9ca3af;margin-top:40px;">3~5초 후 자동으로 사라짐</div>
+      </div>
+    </div>
+    <div class="card-body">
+      <h3>A. 토스트 팝업</h3>
+      <p>우상단에 잠깐 뜨고 자동으로 사라짐. 간단하고 익숙한 패턴.</p>
+    </div>
+  </div>
+
+  <div class="card" data-choice="b" onclick="toggleSelect(this)">
+    <div class="card-image">
+      <div style="font-family:sans-serif;font-size:12px;background:#f4f6f5;border-radius:8px;overflow:hidden;border:1px solid #ddd;">
+        <!-- Header with bell -->
+        <div style="background:#1a1a2e;color:#fff;padding:8px 12px;display:flex;align-items:center;justify-content:space-between;font-size:11px;">
+          <span>ThunderCrew Admin</span>
+          <div style="position:relative;">
+            <span style="font-size:14px;cursor:pointer;">🔔</span>
+            <span style="position:absolute;top:-4px;right:-4px;background:#ef4444;color:#fff;border-radius:999px;font-size:8px;padding:1px 4px;">2</span>
+          </div>
+        </div>
+        <!-- Dropdown -->
+        <div style="background:#fff;border:1px solid #e5e7eb;border-radius:6px;margin:6px;font-size:10px;overflow:hidden;box-shadow:0 4px 12px rgba(0,0,0,.1);">
+          <div style="padding:6px 10px;font-weight:700;border-bottom:1px solid #f3f4f6;font-size:11px;">알림</div>
+          <div style="padding:8px 10px;border-bottom:1px solid #f3f4f6;background:#eff6ff;">
+            <div style="font-weight:600;margin-bottom:2px;">🔔 콜 발송됨</div>
+            <div style="color:#6b7280;">서울12가3456 → 강남구 역삼동</div>
+            <div style="color:#9ca3af;font-size:9px;">방금 전</div>
+          </div>
+          <div style="padding:8px 10px;background:#eff6ff;">
+            <div style="font-weight:600;margin-bottom:2px;">🔔 콜 발송됨</div>
+            <div style="color:#6b7280;">서울34나5678 → 서초구 반포동</div>
+            <div style="color:#9ca3af;font-size:9px;">1분 전</div>
+          </div>
+        </div>
+      </div>
+    </div>
+    <div class="card-body">
+      <h3>B. 헤더 벨 아이콘 + 드롭다운</h3>
+      <p>상단에 🔔 아이콘, 발송할 때마다 뱃지 숫자 증가. 클릭하면 이력 확인.</p>
+    </div>
+  </div>
+
+  <div class="card" data-choice="c" onclick="toggleSelect(this)">
+    <div class="card-image">
+      <div style="font-family:sans-serif;font-size:12px;background:#f4f6f5;border-radius:8px;overflow:hidden;border:1px solid #ddd;">
+        <div style="background:#1a1a2e;color:#fff;padding:8px 12px;font-size:11px;">ThunderCrew Admin</div>
+        <div style="padding:8px;display:grid;grid-template-columns:1fr 1fr;gap:6px;">
+          <!-- Vehicle list -->
+          <div style="background:#fff;border-radius:6px;padding:8px;border:1px solid #e5e7eb;font-size:10px;">
+            <div style="font-weight:700;margin-bottom:6px;">차량</div>
+            <div style="background:#eff6ff;border-radius:4px;padding:4px 6px;border-left:3px solid #3b82f6;margin-bottom:3px;">서울12가3456</div>
+            <div style="padding:4px 6px;color:#6b7280;">서울34나5678</div>
+          </div>
+          <!-- Schedule panel with notification strip -->
+          <div style="background:#fff;border-radius:6px;padding:8px;border:1px solid #e5e7eb;font-size:10px;">
+            <div style="font-weight:700;margin-bottom:4px;color:#3b82f6;">클리닝 일정</div>
+            <!-- Notification strip -->
+            <div style="background:#dcfce7;border:1px solid #86efac;border-radius:4px;padding:5px 7px;margin-bottom:6px;font-size:9px;">
+              <div style="font-weight:700;color:#15803d;margin-bottom:1px;">✓ 콜 발송 완료</div>
+              <div style="color:#166534;">서울12가3456 · 10:00 강남구 역삼동</div>
+            </div>
+            <div style="background:#f9fafb;border-radius:4px;padding:5px 6px;font-size:9px;border:1px solid #e5e7eb;">
+              <div style="color:#6b7280;margin-bottom:1px;">대기 중</div>
+              <div>05/27 14:00 서초구 반포동</div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+    <div class="card-body">
+      <h3>C. 패널 내 인라인 상태 표시</h3>
+      <p>별도 팝업 없이, 일정 패널 안에서 "콜 발송 완료" 상태가 바뀌어 표시.</p>
+    </div>
+  </div>
+
+</div>
