+++ development/front-admin-web/.superpowers/brainstorm/1755-1779775818/content/waiting-1.html
@@ -1,0 +1,3 @@
+<div style="display:flex;align-items:center;justify-content:center;min-height:60vh">
+  <p class="subtitle">Continuing in terminal...</p>
+</div>
