+++ development/front-admin-web/.superpowers/brainstorm/1755-1779775818/state/server-info
@@ -1,0 +1,1 @@
+{"type":"server-started","port":57761,"host":"127.0.0.1","url_host":"localhost","url":"http://localhost:57761","screen_dir":"C:\\Users\\user\\repositories\\clever\\thundercrew-domain\\development\\front-admin-web\\.superpowers\\brainstorm\\1755-1779775818\\content","state_dir":"C:\\Users\\user\\repositories\\clever\\thundercrew-domain\\development\\front-admin-web\\.superpowers\\brainstorm\\1755-1779775818\\state"}
