--- development/front-admin-web/components/overview/OverviewMapBanner.tsx
@@ -1,249 +1,0 @@
-"use client";
-
-import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-
-import { MapShell } from "@/components/dashboard/MapShell";
-import { VehicleDetailDialog, type VehicleDetailRow } from "@/components/management/VehicleDetailDialog";
-import { useFleetSimulation } from "@/components/overview/FleetSimulationContext";
-import { OverviewMapSearch, type OverviewMapSearchMatch } from "@/components/overview/OverviewMapSearch";
-import { useSimulatedBikePins } from "@/components/overview/use-simulated-bike-pins";
-import { useTrailWaypoints } from "@/components/overview/use-trail-waypoints";
-import { useVehicleFilter } from "@/components/overview/VehicleFilterContext";
-import { ServiceTypeFilterTabs, type ServiceTypeFilter } from "@/components/overview/ServiceTypeFilterTabs";
-import type { InsuranceOption } from "@/components/management/RidersPanel";
-import type {
-  FrontendDashboardBikePin,
-  FrontendDashboardStationPin,
-  FrontendVehicle,
-  ServiceOpsInsuranceItem,
-  ServiceOpsRiderInsurance
-} from "@/lib/services/service-ops-api";
-
-/**
- * 루트 페이지(`/`) 상단에 박히는 글로벌 "지도 보기" 토글 + 지도.
- *
- * 토글 OFF 상태에선 MapShell 을 mount 하지 않아 NCP 지도 SDK 세션이 생기지
- * 않는다 — 운영자가 지도가 필요 없을 때까지 API 미터를 잡지 않도록.
- *
- * 차량 탭 활성 + 필터 적용 상태에선 `VehicleFilterContext` 의
- * `filteredBikeIds` 로 부분 집합만 마커로 노출. 다른 탭에선 VehiclesPanel 이
- * 언마운트되며 context 가 null 로 되돌아가 전체 핀 복귀.
- *
- * 차량 상세 floating panel:
- *   - 행 클릭(VehiclesPanel) 또는 지도 마커 클릭 → `selectedBikeId` 가 context
- *     에 publish 됨
- *   - 지도가 닫혀 있으면 자동으로 켜지고 그 차량 위치로 pan
- *   - VehicleDetailDialog 를 **지도 캔버스 내부** 우상단에 floating 으로
- *     렌더링 (옛 /monitoring 의 BikeDetailPanel 패턴). 페이지의 다른 영역
- *     (표 / 탭) 과 상호작용을 막지 않는 비-모달
- */
-export function OverviewMapBanner({
-  bikePins,
-  stationPins,
-  vehicles,
-  bikeActiveRiderById,
-  riderInfoById,
-  riderAllInsurancesByRiderId,
-  insuranceItemById,
-  insuranceOptions
-}: {
-  bikePins: ReadonlyArray<FrontendDashboardBikePin>;
-  stationPins: ReadonlyArray<FrontendDashboardStationPin>;
-  /** 차량 단위 lookup. selectedBikeId 로 FrontendVehicle 을 derive 하는 데 쓴다. */
-  vehicles: ReadonlyArray<FrontendVehicle>;
-  /** bikeId → 활성 라이더 id. detail panel 의 "이름/연락처" 컬럼 lookup 시작점. */
-  bikeActiveRiderById?: Map<string, string>;
-  /** riderId → {name, phone}. */
-  riderInfoById?: Map<string, { name: string; phone: string }>;
-  /** riderId → 활성 rider_insurance 전체 목록. VehicleDetailDialog 보험 편집에 사용. */
-  riderAllInsurancesByRiderId?: Map<string, ServiceOpsRiderInsurance[]>;
-  /** insurance_item id → item. PRIMARY/ADDON 분류 lookup. */
-  insuranceItemById?: Map<string, ServiceOpsInsuranceItem>;
-  /** 보험 상품 선택지. VehicleDetailDialog 보험 편집 select/checkbox 에 사용. */
-  insuranceOptions?: ReadonlyArray<InsuranceOption>;
-}) {
-  const [open, setOpen] = useState(false);
-  const [serviceTypeFilter, setServiceTypeFilter] = useState<ServiceTypeFilter>("ALL");
-  const { filteredBikeIds, selectedBikeId, setSelectedBikeId, setFullscreenMapOpen } = useVehicleFilter();
-  const { seedBikePins } = useFleetSimulation();
-
-  const overlaidBikePins = useSimulatedBikePins(bikePins);
-  const trailWaypoints = useTrailWaypoints(selectedBikeId);
-
-  useEffect(() => {
-    seedBikePins(bikePins);
-  }, [bikePins, seedBikePins]);
-
-  // 검색 결과 클릭이 박는 즉시 팬 좌표. selectedBikeId 기반 자동 팬과 별도
-  // 채널 — BSS 결과는 selectedBikeId 를 안 건드리고 이 override 만 갱신한다.
-  // 매 set 마다 새 객체를 만들어 MapShell 의 targetLocation effect 가 재발화.
-  const [searchOverride, setSearchOverride] = useState<{ lat: number; lng: number } | null>(null);
-
-  const handleSearchSelect = useCallback(
-    (match: OverviewMapSearchMatch) => {
-      // 결과가 클릭되면 무조건 지도부터 켠다 — closed 상태에서 검색만 해도
-      // 사용자가 지도 토글을 따로 누를 필요 없이 즉시 위치 확인 가능.
-      setOpen(true);
-      setSearchOverride({ lat: match.latitude, lng: match.longitude });
-      if (match.kind === "bike" || match.kind === "rider") {
-        setSelectedBikeId(match.bikeId);
-      }
-      // station 종류는 selectedBikeId 를 건드리지 않는다 — BSS 상세 패널을
-      // 두지 않기로 한 스펙 결정. 지도 팬만으로 충분.
-    },
-    [setSelectedBikeId]
-  );
-
-  const bikePinById = useMemo(() => {
-    const map = new Map<string, FrontendDashboardBikePin>();
-    for (const pin of overlaidBikePins) map.set(pin.bikeId, pin);
-    return map;
-  }, [overlaidBikePins]);
-
-  const vehicleById = useMemo(() => {
-    const map = new Map<string, FrontendVehicle>();
-    for (const vehicle of vehicles) {
-      const key = vehicle.id ?? vehicle.slug;
-      if (key) map.set(key, vehicle);
-    }
-    return map;
-  }, [vehicles]);
-
-  const effectiveBikePins = useMemo(() => {
-    let pins = filteredBikeIds === null ? overlaidBikePins : overlaidBikePins.filter((pin) => filteredBikeIds.has(pin.bikeId));
-    if (serviceTypeFilter !== "ALL") {
-      const vehicleServiceType = (bikeId: string) =>
-        vehicleById.get(bikeId)?.serviceType ?? "DELIVERY";
-      pins = pins.filter((pin) => vehicleServiceType(pin.bikeId) === serviceTypeFilter);
-    }
-    return pins;
-  }, [overlaidBikePins, filteredBikeIds, serviceTypeFilter, vehicleById]);
-
-  // 행 클릭이든 마커 클릭이든 selectedBikeId 가 새 값으로 바뀌면 지도를 자동으로 연다.
-  //
-  // `open` 을 deps 에 넣으면 "마커 클릭 → 지도 열림 → 운영자가 토글 끔 → open=false
-  // → effect 재발화 → 다시 켜짐" 루프가 생겨 토글이 동작하지 않는다.
-  // 대신 selectedBikeId 의 이전 값을 ref 로 추적해 값이 바뀌는 시점에만 발화.
-  const prevSelectedBikeIdRef = useRef<string | null>(null);
-  useEffect(() => {
-    if (typeof window === "undefined") return;
-    const prev = prevSelectedBikeIdRef.current;
-    prevSelectedBikeIdRef.current = selectedBikeId;
-    // 새 차량을 선택할 때만 지도를 연다. 같은 selectedBikeId 가 유지되는 동안
-    // open 이 바뀌어도 재발화하지 않아 운영자가 지도를 닫은 상태를 유지 가능.
-    if (selectedBikeId && selectedBikeId !== prev) {
-      const handle = window.requestAnimationFrame(() => setOpen(true));
-      return () => window.cancelAnimationFrame(handle);
-    }
-  }, [selectedBikeId]);
-
-  const targetLocation = useMemo(() => {
-    if (searchOverride) {
-      return { lat: searchOverride.lat, lng: searchOverride.lng };
-    }
-    if (!selectedBikeId) return null;
-    const pin = bikePinById.get(selectedBikeId);
-    if (!pin) return null;
-    // 매 호출마다 새 객체 — MapShell 의 targetLocation effect 가 같은 차량
-    // 두 번 클릭에도 재발화.
-    return { lat: pin.latitude, lng: pin.longitude };
-  }, [searchOverride, selectedBikeId, bikePinById]);
-
-  // 검색 override 는 그 클릭 한 번에만 의미가 있다. 다음에 selectedBikeId
-  // 가 다른 차량으로 바뀌면 (예: 표 행 클릭, 다른 검색 결과) follow 흐름에
-  // 다시 양보하도록 override 를 비운다. override 가 이미 null 이면 마운트
-  // 시점 / station 클릭 직후 등에서 불필요한 rAF 스케줄링 + MapShell 재팬을
-  // 만들지 않도록 short-circuit.
-  useEffect(() => {
-    if (typeof window === "undefined") return;
-    if (!searchOverride) return;
-    const handle = window.requestAnimationFrame(() => setSearchOverride(null));
-    return () => window.cancelAnimationFrame(handle);
-  }, [selectedBikeId, searchOverride]);
-
-  // VehicleDetailDialog 에 넘길 row 데이터. selectedBikeId 가 잡힌 순간 lookup
-  // — vehicle 자체가 없으면(예: 옛 ID 잔존) panel 도 안 뜬다.
-  const detailRow: VehicleDetailRow | null = useMemo(() => {
-    if (!selectedBikeId) return null;
-    const vehicle = vehicleById.get(selectedBikeId);
-    if (!vehicle) return null;
-    const riderId = bikeActiveRiderById?.get(selectedBikeId) ?? null;
-    const riderInfo = riderId ? riderInfoById?.get(riderId) ?? null : null;
-    // 보험 데이터: riderId 로 전체 활성 보험 목록 조회 후 PRIMARY/ADDON 분류.
-    const riderInsurances = riderId ? (riderAllInsurancesByRiderId?.get(riderId) ?? []) : [];
-    const primaryIns = riderInsurances.find((ins) => {
-      const item = insuranceItemById?.get(ins.insuranceItemId);
-      return !item?.category || item.category === "PRIMARY";
-    }) ?? (insuranceItemById ? null : riderInsurances[0] ?? null);
-    const addonInsurances = insuranceItemById
-      ? riderInsurances.filter((ins) => insuranceItemById.get(ins.insuranceItemId)?.category === "ADDON")
-      : [];
-    return {
-      vehicle,
-      riderName: riderInfo?.name ?? null,
-      riderPhone: riderInfo?.phone ?? null,
-      riderId,
-      currentPrimaryInsuranceId: primaryIns?.id ?? null,
-      currentPrimaryInsuranceItemId: primaryIns?.insuranceItemId ?? null,
-      addonInsurances: addonInsurances.map((ins) => ({ id: ins.id, itemId: ins.insuranceItemId }))
-    };
-  }, [selectedBikeId, vehicleById, bikeActiveRiderById, riderInfoById, riderAllInsurancesByRiderId, insuranceItemById]);
-
-  const totalLabel =
-    filteredBikeIds === null
-      ? `${bikePins.length}대 차량`
-      : `${effectiveBikePins.length} / ${bikePins.length}대 차량 (필터)`;
-
-  return (
-    <section className="overview-map-section" aria-label="지도 보기">
-      <div className="overview-map-toggle-row">
-        <label className="overview-map-toggle">
-          <input
-            type="checkbox"
-            checked={open}
-            onChange={(event) => setOpen(event.target.checked)}
-          />
-          <span>지도 보기</span>
-        </label>
-        <ServiceTypeFilterTabs value={serviceTypeFilter} onChange={setServiceTypeFilter} />
-        <OverviewMapSearch
-          bikePins={bikePins}
-          stationPins={stationPins}
-          bikeActiveRiderById={bikeActiveRiderById ?? new Map()}
-          riderInfoById={riderInfoById ?? new Map()}
-          onSelect={handleSearchSelect}
-        />
-        <button
-          type="button"
-          className="overview-map-fullscreen-button"
-          onClick={() => setFullscreenMapOpen(true)}
-          title="전체화면 지도 보기"
-        >
-          ⛶ 전체화면
-        </button>
-        <span className="overview-map-toggle-hint">
-          {totalLabel} · {stationPins.length}개 BSS
-        </span>
-      </div>
-      {open ? (
-        <div className="overview-map-canvas">
-          <MapShell
-            bikePins={[...effectiveBikePins]}
-            stationPins={[...stationPins]}
-            targetLocation={targetLocation}
-            onBikeSelect={setSelectedBikeId}
-            trailWaypoints={trailWaypoints}
-          />
-          {/* 지도 위 floating 상세 패널 — 캔버스 내부 우상단. ESC / 닫기 누르면
-              context 의 selectedBikeId 만 해제, 지도는 그대로 둔다. */}
-          <VehicleDetailDialog
-            key={detailRow ? (detailRow.vehicle.id ?? detailRow.vehicle.slug) : "none"}
-            row={detailRow}
-            insuranceOptions={insuranceOptions ?? []}
-            onClose={() => setSelectedBikeId(null)}
-          />
-        </div>
-      ) : null}
-    </section>
-  );
-}
